Creación de login.html y registro.html

+++ src/main/java/pe/com/cibertec/repository/ProductoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import pe.com.cibertec.entity.ProductoEntity;
 
-public interface ProductoRepository extends JpaRepository<ProductoEntity, Long> {
+public interface ProductoRepository extends JpaRepository<ProductoEntity, Integer> {
     // Aquí ya tienes todos los métodos básicos: findAll(), findById(), save(), delete(), etc.
 }
